Refactor public_tokens to publictokens
